TokenSendingController is updated
