+++ blockchain_test_API/Program.cs
@@ -4,7 +4,6 @@ using blockchain_test_API.Modules;
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
-
 builder.Services.AddCors();
 builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
