+++ blockchain_test_API/blockchain_test_API.csproj
@@ -18,4 +18,8 @@
     </Reference>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Modules\Interfaces\" />
+  </ItemGroup>
+
 </Project>
